Added sketching / button functionality
Added functionality for being able to draw on the grid. Functionality
for choosing colors from the color picker input, where you can choose
the color you want to draw with on the grid. Added functionality for
toggling the grid lines on or off. Added functionality for changing
the grid size from 1x1 to 100x100 cells. Added functionality for
toggling rainbow colors on or off, which draws a random colors. Added
functionality for clearing the grid, which well, clears the grid colored
cells. Added functionality for toggling progressive darkening on or off.
Progressive darkening, causes the colors to be initially 0.1 alpha when
drawn, but gets progressively more visible as it will be incremented
by 0.1 alpha each time you draw over the same cell.